fix: add shadow color to themes
