+++ js/packages/components/chat/message/AudioMessage.tsx
@@ -184,7 +184,7 @@ export const AudioMessage: React.FC<{
 					isHighlight && {
 						borderColor: colors['background-header'],
 						borderWidth: 1,
-						shadowColor: colors['background-header'],
+						shadowColor: colors.shadow,
 						shadowOffset: {
 							width: 0,
 							height: 8,
