+++ js/packages/components/modals/ManageGroupInvitation.tsx
@@ -97,7 +97,7 @@ export const ManageGroupInvitation: React.FC<{
 					<MultiMemberAvatar
 						publicKey={publicKey}
 						fallbackNameSeed={displayName}
-						style={[border.shadow.big, row.center] as any}
+						style={[border.shadow.big, row.center, { shadowColor: colors.shadow }] as any}
 						size={100}
 					/>
 				</View>
@@ -209,7 +209,12 @@ export const ManageGroupInvitation: React.FC<{
 					row.item.justify,
 					column.justify,
 					_styles.closeRequest,
-					{ position: 'absolute', bottom: '2%', backgroundColor: colors['main-background'] },
+					{
+						position: 'absolute',
+						bottom: '2%',
+						backgroundColor: colors['main-background'],
+						shadowColor: colors.shadow,
+					},
 				]}
 				onPress={navigation.goBack}
 			>
