+++ js/packages/components/settings/DeleteAccount.tsx
@@ -65,7 +65,7 @@ const DeleteAccountHeader: React.FC<{ title: string }> = ({ title }) => {
 					row.item.justify,
 					column.justify,
 					_styles.header,
-					{ backgroundColor: colors['main-background'] },
+					{ backgroundColor: colors['main-background'], shadowColor: colors.shadow },
 				]}
 			>
 				<Icon
